web_backEnd/src/ShopWeb.Tests/Service/OrderServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using ShopWeb.Core.src.Common;
using ShopWeb.Core.src.Entity;
using ShopWeb.Core.src.RepoAbstract;
using ShopWeb.Service.src.DTOs;
using ShopWeb.Service.src.Services;
using Xunit;

namespace ShopWeb.Service.Tests
{
    public class OrderServiceTests
    {
        private readonly OrderService _orderService;
        private readonly Mock<IOrderRepository> _orderRepositoryMock;
        private readonly Mock<IOrderLineItemsRepository> _orderLineItemsRepositoryMock;
        private readonly Mock<IMapper> _mapperMock;
        private readonly Mock<IAddressRepository> _addressRepositoryMock;

        public OrderServiceTests()
        {
            _orderRepositoryMock = new Mock<IOrderRepository>();
            _orderLineItemsRepositoryMock = new Mock<IOrderLineItemsRepository>();
            _mapperMock = new Mock<IMapper>();
            _addressRepositoryMock = new Mock<IAddressRepository>();

            _orderService = new OrderService(
                _orderRepositoryMock.Object,
                _orderLineItemsRepositoryMock.Object,
                _mapperMock.Object,
                _addressRepositoryMock.Object
            );
        }

        [Fact]
        public async Task CreateOrderAsync_ValidUserIdAndOrderDto_ReturnsCreatedOrder()
        {
            // Arrange
            var userId = Guid.NewGuid();
            var orderCreateDto = new OrderCreateDto
            {
                // Set up order create DTO
            };
            var order = new Order
            {
                // Set up order
            };

            _mapperMock.Setup(x => x.Map<Order>(orderCreateDto)).Returns(order);
            _orderRepositoryMock.Setup(x => x.CreateNewOrderAsync(order)).ReturnsAsync(order);
            _mapperMock.Setup(x => x.Map<OrderReadDto>(order)).Returns(new OrderReadDto());

            // Act
            var result = await _orderService.CreateOrderAsync(userId, orderCreateDto);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task DeleteOrderByIdAsync_ValidOrderId_ReturnsTrue()
        {
            // Arrange
            var orderId = Guid.NewGuid();

            _orderRepositoryMock.Setup(x => x.DeleteOrderByIdAsync(orderId)).ReturnsAsync(true);

            // Act
            var result = await _orderService.DeleteOrderByIdAsync(orderId);

            // Assert
            Assert.True(result);
        }

        [Fact]
        public async Task GetAllOrdersAsync_ReturnsAllOrders()
        {
            // Arrange
            var orders = new List<Order>
            {
                // Set up list of orders
            };
            _orderRepositoryMock.Setup(x => x.GetAllOrdersAsync()).ReturnsAsync(orders);
            _mapperMock.Setup(x => x.Map<IEnumerable<OrderReadDto>>(orders)).Returns(new List<OrderReadDto>());

            // Act
            var result = await _orderService.GetAllOrdersAsync();

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetOrderByIdAsync_ValidOrderId_ReturnsOrder()
        {
            // Arrange
            var orderId = Guid.NewGuid();
            var order = new Order
            {
                // Set up order
            };
            _orderRepositoryMock.Setup(x => x.GetOrderByIdAsync(orderId)).ReturnsAsync(order);
            _mapperMock.Setup(x => x.Map<OrderReadDto>(order)).Returns(new OrderReadDto());

            // Act
            var result = await _orderService.GetOrderByIdAsync(orderId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetOrdersByUserIdAsync_ValidUserId_ReturnsOrders()
        {
            // Arrange
            var userId = Guid.NewGuid();
            var orders = new List<Order>
            {
                // Set up list of orders
            };
            _orderRepositoryMock.Setup(x => x.GetOrdersByUserIdAsync(userId)).ReturnsAsync(orders);
            _mapperMock.Setup(x => x.Map<IEnumerable<OrderReadDto>>(orders)).Returns(new List<OrderReadDto>());

            // Act
            var result = await _orderService.GetOrdersByUserIdAsync(userId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }
    }
}
